moved SortedMap & SortedMapK to arrow-core-data
