+++ modules/core/arrow-core-data/src/main/kotlin/arrow/core/SortedMap.kt
@@ -1,3 +1,3 @@
-package arrow.data
+package arrow.core
 
 typealias SortedMap<K, V> = java.util.SortedMap<K, V>
