+++ modules/core/arrow-core-data/src/main/kotlin/arrow/core/SortedMapK.kt
@@ -1,12 +1,6 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind
-import arrow.core.Eval
-import arrow.core.Option
-import arrow.core.Tuple2
-import arrow.core.identity
-import arrow.core.iterateRight
-import arrow.core.k as mk
 import arrow.higherkind
 import arrow.typeclasses.Applicative
 
@@ -50,7 +44,7 @@ data class SortedMapK<A : Comparable<A>, B>(private val map: SortedMap<A, B>) : 
 
   fun <G, C> traverse(GA: Applicative<G>, f: (B) -> Kind<G, C>): Kind<G, SortedMapK<A, C>> = GA.run {
     map.iterator().iterateRight(Eval.always { just(sortedMapOf<A, C>().k()) }) { kv, lbuf ->
-      f(kv.value).map2Eval(lbuf) { (mapOf(kv.key to it.a).mk() + it.b).toSortedMap().k() }
+      f(kv.value).map2Eval(lbuf) { (mapOf(kv.key to it.a).k() + it.b).toSortedMap().k() }
     }.value()
   }
