+++ modules/core/arrow-test/src/main/kotlin/arrow/test/generators/Generators.kt
@@ -23,16 +23,14 @@ import arrow.core.Tuple8
 import arrow.core.Tuple9
 import arrow.core.fix
 import arrow.core.toOption
-import arrow.core.extensions.option.functor.map
 import arrow.core.extensions.option.functor.functor
 import arrow.core.ListK
 import arrow.core.MapK
 import arrow.core.NonEmptyList
 import arrow.core.SequenceK
 import arrow.core.SetK
-import arrow.data.SortedMapK
+import arrow.core.SortedMapK
 import arrow.core.Validated
-import arrow.data.k
 import arrow.core.k
 import arrow.recursion.Algebra
 import arrow.recursion.Coalgebra
