+++ modules/docs/arrow-docs/docs/docs/arrow/data/sortedmapk/README.md
@@ -15,7 +15,6 @@ beginner
 
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.data.*
 import arrow.core.*
 
 DataType(SortedMapK::class).tcMarkdownList()
